[powerbug/scan.sh] added basic checking for consensus, some minor improvements

+++ problems/2D_RIEMANN/powerbug/scan.sh
@@ -6,13 +6,15 @@
 NTHR=$( lscpu -p | grep -v '^#' | wc -l )
 
 # If there is mprime  running for warm-up, stop it.
-killall -STOP mprime
+MPRIME=1
+killall -STOP mprime 2> /dev/null || MPRIME=0
 
 SHOW_TEMP=.__st__
 # A "daemon" for reporting temperatures. Adjust the grep filter to particular machine.
 touch $SHOW_TEMP
 sleep 5 && while [ -e $SHOW_TEMP ] ; do
     ( cat $SHOW_TEMP; sensors | grep -E '(Tctl|TSI0_TEMP)' ) | tr '\n' ' ' ; echo
+    #( cat $SHOW_TEMP; sensors | grep -E '(Package)' ) | tr '\n' ' ' ; echo
     sleep 5
 done &
 
@@ -27,9 +29,11 @@ done
 # Scan for most energy-consuming setup parameters.
 # It seems that n_d has the biggest impact. Its optimal size may change with size and organization of the CPU cache.
 
-# For a Ryzen 7 2700 maximum power draw occurs at bs=300 and RTVD solver
+# Maximum power draw depends on CPU model and problem configuration:
+# Ryzen 7 2700: bs=300 solver=RTVD
+# i5-11600: bs=116 solver=RTVD
 
-for bs in 240 300 360 ; do
+for bs in 116 300 ; do
     k=1
     for s in Riemann RTVD ; do
 	nd=$(( $bs * $k ))
@@ -38,13 +42,37 @@ for bs in 240 300 360 ; do
 	# Parallel Piernik runs do spend some time on communication, so we choose flooding until we implement shared-memory :)
 	for d in `seq $NTHR` ; do
 	    cd $d
+	    rm *log 2> /dev/null
 	    ./piernik -n '&BASE_DOMAIN n_d = 2*'$nd', 1/ &AMR bsize = 2*'$bs', 1/ &NUMERICAL_SETUP solver_str="'$s'"' > /dev/null &
 	    cd - > /dev/null
 	done
-	# Run each Piernik configuration for 120 seconds.
-	sleep 120 && killall piernik
+
+	# Check if all instances are reaching the same values.
+        OSTEP=0
+	while [ -e $SHOW_TEMP ] ; do
+	    [ -e 1/*log ] && \
+		STEP=$( tail -n 1 [1-9]*/*log | \
+			    grep nstep | \
+			    sort -n -k 5 | \
+			    head -n 1 | \
+			    awk '{print $5}' )
+            [ ${STEP:-0} -gt 0 ] && \
+                if [ $STEP != $OSTEP ] ; then
+                    OSTEP=$STEP
+                    grep "nstep = *$STEP" [1-9]*/*log | \
+			sort -n | \
+			column -t | \
+			awk '{if (NR==1) {dt=$9; t=$12; print} else if (dt != $9 || t != $12) print "Differs: ", $0}'
+                fi
+            sleep 3
+        done &
+
+        # ToDo also check if all $NTHR are alive and not stuck.
+
+	# Run each Piernik configuration for 60 seconds.
+	sleep 60 && killall piernik
 	echo "kill $nd $k"
-	killall -9 piernik
+	sleep 1 && killall -9 piernik
 	echo
     done
 done
@@ -52,7 +80,7 @@ done
 rm $SHOW_TEMP
 
 # Continue mprime, if it was there.
-killall -CONT mprime
+[ $MPRIME == 1 ]  && killall -CONT mprime
 
 # This may help to visualize CPU temperature changes.
 # gnuplot
